Fix chart data set and renderer issues

+++ Source/Charts/Renderers/BarLineScatterCandleBubbleRenderer.swift
@@ -66,7 +66,7 @@ open class BarLineScatterCandleBubbleRenderer: NSObject, DataRenderer
     open func isDrawingValuesAllowed(dataProvider: ChartDataProvider?) -> Bool
     {
         guard let data = dataProvider?.data else { return false }
-        return data.entryCount < Int(CGFloat(dataProvider?.maxVisibleCount ?? 0) * viewPortHandler.scaleX)
+        return CGFloat(data.entryCount) < (CGFloat(dataProvider?.maxVisibleCount ?? 0) * viewPortHandler.scaleX)
     }
 
     /// Class representing the bounds of the current viewport in terms of indices in the values array of a DataSet.
